refactor: split LanguageAvailabilityPage into smaller components
- Extract LanguageAvailabilityRow, DraftLanguageRow, and AddLanguageModal components
- Create hooks for mutations, drag-and-drop, and filters
- Remove debug console.log statements
- Reduce main page component from 1,249 to 508 lines
- Fix TypeScript errors and ensure all CI checks pass

ref EL-118
